backend/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
awscli==1.22.101
botocore==1.24.46
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.4
django-cors-headers==3.11.0
django-model-utils==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
docutils==0.15.2
drf-yasg==1.20.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
msgpack==1.0.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
supervisor==4.2.4
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
